Minor bug fixes
Prior to this, you could only buy <max_tickets> at a time, but you could run the command numerous times to bypass it. You can no longer do this, as the limit is now on the amount of tickets the player *owns* rather than the amount of tickets the player is buying with the /ql ticket buy <amount> command.